requirements.txt
torch
pillow
ray
opencv-python
torchvision